Add Z80 test harness, fix register bug in AI
Wrote a Python Z80 emulator (just the opcodes we actually use)
to test the chess code without needing a full emulator. Found a
real bug in the process: think_scan was doing PUSH AF then POP DE
to get the square number back, but that puts A in D not E. The
piece type was getting clobbered. Fixed with POP AF / LD E,A.